+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Remove extractor's `id` field, use `identifier` field instead
 - Rework document polymorphism, the model used to deserialize document response is now automatically chosen based on the document's extractor
+- Make Document.Meta.confirmedBy nullable
 
 ### Removed
 - Remove extractor's `id` field